backend/requirements.txt
pymongo
flask   
python-dotenv
flask-cors
numpy
scikit-learn
